pyproject.toml
[project]
name = "exp-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "encourage-rag==0.3.7.4",
    "hydra-core>=1.3.2",
    "vllm>=0.6.2",
    "mlflow>=2.4.0",
    "datasets==3.5.0",
    "numpy>=1.26.4",
    "google-cloud-storage==3.1.0",
    "chromadb==1.0.10",
]
[build-system]
requires = ["setuptools", "setuptools-scm", "torch"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["exp*"]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "W", "I", "D", "A", "N", "B", "SIM", "C4", "TID"]
lint.ignore = [
    "D100", # Missing docstring in public module
    "E741", # Ambiguous variable name
    "D213", # Multi-line docstring summary should start at the second line
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D401", # First line should be in imperative mood; try rephrasing
    "D407", # Missing dashed underline after section
    "D406", # Section name should end with a newline
    "D106", # Missing docstring in public nested class
    "B006", # Do not use mutable data structures for argument defaults
    "B008", # Do not perform function calls in argument defaults
    "B905", # `zip()` without an explicit `strict=` parameter
]
